backend-pwa/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
langchain==0.0.335
langchain-openai==0.0.2.post1
httpx==0.25.0
python-multipart==0.0.6
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
